requirements.txt
pandas
scikit-learn
streamlit
matplotlib
seaborn
plotly
nbformat
opencv-python
Pillow
glob2
python-dotenv
fastapi
uvicorn
requests
ipykernel
psutil